% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{c}
\alias{c}
\alias{c.gPair}
\title{c.gPair}
\usage{
\method{c}{gPair}(...)
}
\arguments{
\item{gPair}{object}
}
\value{
a new concatenated gPair Object
}
\description{
Concatenates gPair Objects
}
\author{
Marcin Imielinski
}
