% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{cocount}
\alias{cocount}
\title{cocount}
\usage{
cocount(
  events,
  bins = disjoin(events),
  by = names(values(events))[1],
  weight = NULL,
  frac = FALSE,
  full = FALSE,
  fill = 0,
  na.rm = TRUE
)
}
\arguments{
\item{events}{GRanges of events}

\item{bins}{bins to count to, default is disjoin(events)}

\item{by}{column of events GRanges that allows co-counting}

\item{frac}{whether to count fractional overlap of event with bin as 1 (frac == FALSE) or based on width}
}
\description{
gMatrix instantiator creates contact "counts" of "events", which are GRanges grouped into
pairs / triples / groups (e.g. paired end reads, linked readsa, SPRITE contacts) via "by"
column.
}
