% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{gM}
\alias{gM}
\title{gM}
\usage{
gM(
  gr = NULL,
  dat = NULL,
  outer = NULL,
  field = NULL,
  full = FALSE,
  fill = 0,
  agg.fun = sum,
  na.rm = TRUE,
  lower.tri = FALSE
)
}
\arguments{
\item{gr}{GRanges around which to build a gMatrix or gPair object to sum}

\item{dat}{data.table of $i and $j indexing gr and field $value OR a vector valued expression involving terms i and j that will be applied to expand.grid}

\item{outer}{character scalar field of gr that will be used to generate an "outer product" of gr with itself, which given vector x for numerics will make each gmatrix entry equal to x_i*x_j and for character vector or factor fields will make each vector equal to x_i == x_j}

\item{field}{if gr is a gPair then field can be a character that specifies a gPair metadata numeric or integer metadata field that will be used as weights to sum the gPair objects}

\item{full}{logical flag whether to explicitly "fill" missing entries in provided data.table (default = FALSE)}

\item{fill}{value with which to fill missing values (default 0)}

\item{agg.fun}{default function with which to aggregate values, this function should take a vector and na.rm = TRUE and return a scalar}

\item{na.rm}{default na.rm argument to agg.fun}
}
\value{
A new gMatrix object
}
\description{
gMatrix instantiator
}
\author{
Marcin Imielinski
}
