% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{gP}
\alias{gP}
\title{gP}
\usage{
gP(gr1 = NULL, gr2 = NULL, meta = NULL)
}
\arguments{
\item{gr1}{GRanges around which to build a gPair, also can be a GRAngesList, in which case all intra-list item pairs are created.}

\item{gr2}{optional second granges same length as the first (but possibly different width), by default is set to gr1}

\item{meta}{data.table of metadata (by default pulled from values(gr1)}
}
\value{
A new gPair object
}
\description{
gPair instantiator, creates a pair with some metadata
}
\author{
Marcin Imielinski
}
