% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\docType{methods}
\name{\%$\%}
\alias{\%$\%}
\alias{\%$\%,gMatrix-method}
\title{annotate metadata GRanges $gr of a gMatrix with those of another}
\usage{
\S4method{\%$\%}{gMatrix}(x, y)
}
\arguments{
\item{x}{\code{gMatrix} object}
}
\value{
gMatrix with GRanges metadata  populated
}
\description{
gm %$% gr

will populate the GRanges metadata of gMatrix gm (i.e. gm$gr)
with the metadata of gr
}
\author{
Marcin Imielinski
}
