% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{grnorm}
\alias{grnorm}
\title{grnorm}
\usage{
grnorm(
  gr,
  mean = NULL,
  sd = NULL,
  full = FALSE,
  fill = 0,
  agg.fun = sum,
  na.rm = TRUE
)
}
\arguments{
\item{gr}{GRanges around which to build a gMatrix}

\item{mean}{mean parameter funtion of i and j}

\item{sd}{scalar sd parameter to rnorm}

\item{full}{logical flag whether to explicitly "fill" missing entries in provided data.table (default = FALSE)}

\item{fill}{value with which to fill missing values (default 0)}

\item{agg.fun}{default function with which to aggregate values, this function should take a vector and na.rm = TRUE and return a scalar}

\item{na.rm}{default na.rm argument to agg.fun}
}
\description{
gMatrix instantiator creates random matrix with normal distribution
}
