% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{grpois}
\alias{grpois}
\title{grpois}
\usage{
grpois(gr, lambda = NULL, full = FALSE, fill = 0, agg.fun = sum, na.rm = TRUE)
}
\arguments{
\item{gr}{GRanges around which to build a gMatrix or gMatrix from which to draw "lambda" parameter}

\item{lambda}{scalar lambda parameter of poisson or gMatrix from which to draw lambda parameter)}

\item{full}{logical flag whether to explicitly "fill" missing entries in provided data.table (default = FALSE)}

\item{fill}{value with which to fill missing values (default 0)}

\item{agg.fun}{default function with which to aggregate values, this function should take a vector and na.rm = TRUE and return a scalar}

\item{na.rm}{default na.rm argument to agg.fun}
}
\description{
gMatrix instantiator creates random matrix with poisson
}
