% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\docType{package}
\name{GxG}
\alias{GxG}
\alias{GxG-package}
\title{GxG}
\description{
Matrices and pairwise features in a GenomicRanges / data.table framework

Copyright (C) 2018 Marcin Imielinski

   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
   (at your option) any later version.

   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.

   You should have received a copy of the GNU General Public License
   along with this program.  If not, see <http://www.gnu.org/licenses/>.

   Github: https://github.com/mskilab/GxG
}
\author{
\strong{Maintainer}: Marcin Imielinski \email{mai9037@med.cornell.edu}

}
