% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{hicpro}
\alias{hicpro}
\title{hicpro}
\usage{
hicpro(mat, bed = NULL)
}
\arguments{
\item{mat}{path to .matrix file}

\item{bed}{path to .bed file}
}
\value{
gMatrix
}
\description{
Instantiates gMatrix from .matrix and .bed Hi-C pro file
}
\author{
Marcin Imielinski
}
\keyword{hicpro}
