% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{intersect}
\alias{intersect}
\title{intersect.gPair}
\usage{
\S4method{intersect}{gPair,gPair}(x, y, pad = 0, ...)
}
\arguments{
\item{x}{a gPair Object}

\item{y}{a gPair Object}
}
\value{
new gPair Object containing the intersection of x and y
}
\description{
Returns a new gPair Object which is the intersection of x and y.
}
\author{
Marcin Imielinski
}
