% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{length}
\alias{length}
\alias{length.gPair}
\title{length.gPair}
\usage{
\method{length}{gPair}(gPair)
}
\arguments{
\item{gPair}{a gPair Object}
}
\value{
the number of pairs in the gPair Object
}
\description{
The number of pairs in this gPair Object
}
\author{
Marcin Imielinski
}
