% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{seqinfo}
\alias{seqinfo}
\title{seqinfo}
\usage{
\S4method{seqinfo}{gMatrix}(x)

\S4method{seqinfo}{gPair}(x)
}
\arguments{
\item{x}{a gPair}
}
\value{
the seqinfo of this gMatrix

the seqinfo of this gPair
}
\description{
seqinfo

seqinfo
}
