% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{seqlengths}
\alias{seqlengths}
\title{seqlengths}
\usage{
\S4method{seqlengths}{gMatrix}(x)

\S4method{seqlengths}{gPair}(x)
}
\arguments{
\item{x}{a gPair}
}
\value{
the seqlengths of this graph

the seqlengths of this graph
}
\description{
seqlengths

seqlengths
}
