% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{setdiff}
\alias{setdiff}
\title{setdiff.gPair}
\usage{
\S4method{setdiff}{gPair,gPair}(x, y)
}
\arguments{
\item{x}{a gPair Object}

\item{y}{a gPair Object}
}
\value{
new gPair Object containing the difference between x and y
}
\description{
Returns a new gPair Object which is the difference between x and y.
}
\author{
Marcin Imielinski
}
