% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{straw}
\alias{straw}
\title{straw}
\usage{
straw(
  hic,
  gr = NULL,
  norm = "NONE",
  type = "BP",
  res = 10000,
  mc.cores = 1,
  colormap = c("white", "red", "black"),
  ...
)
}
\arguments{
\item{hic}{path to .hic file}

\item{gr}{granges to query}

\item{norm}{string specifying normalization to apply ("KR" (default), "NONE", VC")}

\item{type}{"BP" vs "FRAG"}

\item{res}{resolution to query (default 10kb)}

\item{mc.cores}{parallelization to apply for query (useful when length(gr)>2)}
}
\value{
gMatrix
}
\description{
Instantiates gMatrix from .hic object using straw API https://github.com/theaidenlab/straw/tree/master/R
used to extract all of the data in length n query gr (note will do n choose 2 queries since
straw only supports pairwise queries)
}
\author{
Marcin Imielinski
Tuesday, Nov 07, 2017 04:43:52 PM - Julie: adding as.character()
}
\keyword{straw}
