% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{[.gMatrix}
\alias{[.gMatrix}
\title{gMatrix}
\usage{
\method{[}{gMatrix}(obj, i = NULL, j = NULL, with = TRUE, drop = FALSE, full = FALSE)
}
\arguments{
\item{obj}{gMatrix object This is the gMatrix object to be subset}

\item{i}{integer, logical, or expression in gMatrix metadata used to subset gMatrix}
}
\value{
A new gMatrix object that contains only the given id's
}
\description{
Overloads subset operator for gMatrix. Allows subsetting of gMatrix via index.
Also, allows for data.table style queries on gMatrix GRangesList and
corresponding metadata.
}
\author{
Marcin Imielinski
}
