% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{[.gPair}
\alias{[.gPair}
\title{gPair}
\usage{
\method{[}{gPair}(obj, i, j, with = TRUE)
}
\arguments{
\item{obj}{gPair object This is the gPair object to be subset}

\item{i}{integer, logical, or expression in gPair metadata used to subset gPair}
}
\value{
A new gPair object that contains only the given id's
}
\description{
Overloads subset operator for gPair. Allows subsetting of gPair via index.
Also, allows for data.table style queries on gPair GRangesList and
corresponding metadata.
}
\author{
Marcin Imielinski
}
