% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{union}
\alias{union}
\title{union.gPair}
\usage{
\S4method{union}{gPair,gPair}(x, y)
}
\arguments{
\item{x}{a gPair Object}

\item{y}{a gPair Object}
}
\value{
new gPair Object containing the union of x and y
}
\description{
Returns a new gPair Object which is the union of x and y.
}
\author{
Marcin Imielinski
}
