% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{unique}
\alias{unique}
\alias{unique.gPair}
\title{unique.gPair}
\usage{
\method{unique}{gPair}(x, pad = 0, ignore.strand = FALSE)
}
\arguments{
\item{x}{a gPair Object}
}
\value{
new gPair Object containing the union of x and y
}
\description{
Returns the subset of gPair object that it is unique
}
\author{
Marcin Imielinski
}
