% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{homeology.event}
\alias{homeology.event}
\title{homeology.event}
\usage{
homeology.event(
  event,
  pad = 100,
  thresh = 2,
  stride = 1,
  pad2 = 5,
  flip = FALSE,
  mc.cores = 1,
  anchor = TRUE,
  deanchor_gm = TRUE,
  mat = FALSE,
  genome = "~/DB/GATK/human_g1k_v37.fasta",
  bidirected_search = TRUE,
  save_gm = TRUE
)
}
\arguments{
\item{event}{data.table of junctions}

\item{pad}{width around junction breakpoint around which to search for homeology}

\item{thresh}{string distance threshold for calling homeology in a bin}

\item{stride}{distance in bases between consecutive bins in which we will be measuring homeology}

\item{pad2}{number of bases of padding around each sequence position (bin) to use when computing homeology}

\item{flip}{if flip = FALSE, homeology search for -/- and +/+ junctions is done between a sequence and its reverse complement}

\item{mc.cores}{number of cores to use}

\item{anchor}{if TRUE, anchor the junctions to the genome}

\item{deanchor_gm}{if TRUE, deanchor the gMatrix}

\item{mat}{if TRUE, return the gMatrix as a matrix}

\item{genome}{path to .fasta containing genome sequence}

\item{bidirected_search}{if TRUE, add padding on both sides of each breakpoint, if FALSE, add padding only in the direction of the fused side}

\item{save_gm}{if TRUE, save the gMatrix}
}
\description{
homeology.event
}
